+++ mrfilestorage/wrapper_errors.go
@@ -1,11 +1,13 @@
 package mrfilestorage
 
 import (
+	"context"
 	"errors"
 	"io/fs"
 	"os"
 
 	"github.com/mondegor/go-webcore/mrcore"
+	"github.com/mondegor/go-webcore/mrctx"
 )
 
 func (fp *FileProvider) wrapError(err error, skip int) error {
@@ -18,4 +20,13 @@ func (fp *FileProvider) wrapError(err error, skip int) error {
 	}
 
 	return mrcore.FactoryErrInternal.Caller(skip + 1).Wrap(err)
+}
+
+func (fp *FileProvider) debugCmd(ctx context.Context, command, fileName string) {
+	mrctx.Logger(ctx).Debug(
+		"%s: cmd=%s, file=%s",
+		providerName,
+		command,
+		fp.baseDir+fileName,
+	)
 }