+++ mrminio/wrapper_errors.go
@@ -1,8 +1,11 @@
 package mrminio
 
 import (
+	"context"
+
 	"github.com/minio/minio-go/v7"
 	"github.com/mondegor/go-webcore/mrcore"
+	"github.com/mondegor/go-webcore/mrctx"
 )
 
 func (fp *FileProvider) wrapError(err error) error {
@@ -18,4 +21,14 @@ func (fp *FileProvider) wrapError(err error) error {
 	}
 
 	return mrcore.FactoryErrInternal.Caller(1).Wrap(err)
+}
+
+func (fp *FileProvider) debugCmd(ctx context.Context, command, fileName string) {
+	mrctx.Logger(ctx).Debug(
+		"%s: cmd=%s, bucket=%s, file=%s",
+		providerName,
+		command,
+		fp.bucketName,
+		fp.baseDir+fileName,
+	)
 }